Refactor: rename 'prep' to 'parse'.
